# docs/openapi-subscriber.yaml
...
openapi: 3.0.3
info:
  title: Ollie Signal API
  version: 0.1.0
  description: |-
    Subscriber-facing API for the Ollie signal feed.

    `POST /v1/session` is public and turns a Sign in with Apple identity token into an app session token; everything else under /v1 requires that token as a bearer. `/healthz` is public. The MCP endpoint at `/mcp` takes a separately minted MCP token and is not described by this document.

    Nothing here can place, modify, or read an order on any brokerage account. Every response is identical for every subscriber.
components:
  securitySchemes:
    subscriberToken:
      type: http
      scheme: bearer
      description: The app session token from POST /v1/session. Per-subscriber, hashed at rest, revocable. Not accepted at /mcp, which takes an MCP token.
  schemas:
    SessionRequestInput:
      type: object
      properties:
        identity_token:
          type: string
          minLength: 1
          description: The Sign in with Apple identity token, verbatim from ASAuthorization
        invite_code:
          description: Required on the first sign-in while the soft launch is invite-gated
          type: string
          maxLength: 64
      required:
        - identity_token
    AcceptDisclaimerRequestInput:
      type: object
      properties:
        version:
          type: string
          description: The version the client displayed; must match the current one
      required:
        - version
...
    SignalStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
        - expired
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/SignalStatus"
      required:
        - error
      additionalProperties: false
      description: Error response
...
    Subscriber:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        email:
          type: string
          nullable: true
          description: May be a private-relay address or null
        created_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
      required:
        - id
        - email
        - created_at
      additionalProperties: false
    Session:
      type: object
      properties:
        token:
          type: string
          description: The app session bearer token. Store it in the Keychain; it is not shown again.
        subscriber:
          $ref: "#/components/schemas/Subscriber"
      required:
        - token
        - subscriber
      additionalProperties: false
    Disclaimer:
      type: object
      properties:
        version:
          type: string
          description: sha256 of the text. Send it back on accept.
        text:
          type: string
          description: Markdown
      required:
        - version
        - text
      additionalProperties: false
    Onboarding:
      type: object
      properties:
        disclaimer:
          $ref: "#/components/schemas/Disclaimer"
        accepted_current_version:
          type: boolean
          description: Whether this subscriber has an acceptance row for the version above
        has_mcp_token:
          type: boolean
          description: Whether an unrevoked MCP token exists
        mcp_url:
          type: string
          nullable: true
          description: Where to point an agent. Null only when the service has no public URL configured.
      required:
        - disclaimer
        - accepted_current_version
        - has_mcp_token
        - mcp_url
      additionalProperties: false
    DisclaimerAcceptance:
      type: object
      properties:
        version:
          type: string
        accepted_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
      required:
        - version
        - accepted_at
      additionalProperties: false
    McpToken:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        created_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
        last_used_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          nullable: true
        revoked_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          nullable: true
      required:
        - id
        - created_at
        - last_used_at
        - revoked_at
      additionalProperties: false
    McpTokenMinted:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        created_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
        last_used_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          nullable: true
        revoked_at:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          nullable: true
        token:
          type: string
          description: Shown exactly once. Only its hash is stored.
        mcp_url:
          type: string
          nullable: true
      required:
        - id
        - created_at
        - last_used_at
        - revoked_at
        - token
        - mcp_url
      additionalProperties: false
    McpTokenList:
      type: object
      properties:
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/McpToken"
      required:
        - tokens
      additionalProperties: false
...
paths:
...
  /v1/session:
    post:
      operationId: createSession
      summary: Sign in with Apple and receive an app session token
      tags:
        - onboarding
      description: |-
        Verifies the identity token against Apple's keys, then finds or creates the subscriber by Apple subject. A first sign-in needs a valid invite code while the soft launch is gated; a returning subscriber does not.

        The token in the response is the only time it is shown.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequestInput"
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: The request failed schema validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: The identity token did not verify
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: First sign-in without a valid invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/onboarding:
    get:
      operationId: getOnboarding
      summary: Where this subscriber is in onboarding, from rows
      tags:
        - onboarding
      security:
        - subscriberToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
        "401":
          description: Missing or invalid app token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/disclaimer/accept:
    post:
      operationId: acceptDisclaimer
      summary: Record that this subscriber accepted the disclaimer
      tags:
        - onboarding
      description: "The version must be the one currently served. A mismatch means the text changed under the client: re-fetch onboarding, re-display, and accept the new version."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptDisclaimerRequestInput"
      security:
        - subscriberToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisclaimerAcceptance"
        "400":
          description: The request failed schema validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid app token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The version accepted is not the version served
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/mcp-tokens:
    post:
      operationId: mintMcpToken
      summary: Mint an agent-facing MCP token
      tags:
        - tokens
      description: Refused with 409 until the current disclaimer version has been accepted — the credential cannot exist without the consent row predating it. The plaintext is returned once; only its hash is stored. Lost it? Revoke and mint again.
      security:
        - subscriberToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpTokenMinted"
        "401":
          description: Missing or invalid app token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The current disclaimer has not been accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listMcpTokens
      summary: This subscriber's MCP tokens, never the plaintext
      tags:
        - tokens
      security:
        - subscriberToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpTokenList"
        "401":
          description: Missing or invalid app token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/mcp-tokens/{id}:
    delete:
      operationId: revokeMcpToken
      summary: Revoke an MCP token
      tags:
        - tokens
      description: Idempotent. The row remains as history with `revoked_at` set.
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - subscriberToken: []
      responses:
        "204":
          description: Revoked, or already was
        "400":
          description: The request failed schema validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid app token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such token belongs to this subscriber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://ollie-signal.up.railway.app
    description: Railway